requirements.txt
streamlit>=1.28.0
google-genai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
pytest>=7.4.0
dotenv>=1.0.0
